requirements.txt
Flask
flask-cors
google-cloud-firestore